refactor: rename dashboard to web
