+++ web/wasm/events/events.go
