+++ web/wasm/templates/dashboard/main.html
